requirements.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.5
gymnasium>=0.29
stable-baselines3>=2.3.0
torch>=2.3
matplotlib>=3.9
streamlit>=1.37
pydantic>=2.8
pyyaml>=6.0.2
